requirements.txt
submitit
torchmetrics[image]
torchvision
torchdiffeq

opencv-python
seaborn
pydicom
pylic
imageio
albumentations
nibabel